fixed fucking bugs

+++ objects/obj_checkBoxYes/Mouse_7.gml
@@ -1,7 +1,9 @@
 event_inherited();
 
 // yes -> reset save file and restart
-if (file_exists(SAVEFILE)) {
-    file_delete(SAVEFILE);
+if(isPressed){
+	if (file_exists(SAVEFILE)) {
+	    file_delete(SAVEFILE);
+	}
+	game_restart();
 }
-game_restart();
